Formatter uses stdin instead of filepath
In the current version, at least on Nvim and OSX, .clang-format in the
root of the repo was being ignored because clang-format looks for a
config file relative to the source file, not the working directory.

This change allows .clang-format in the root of a workspace to be properly picked up
with the addition of the assume-filename flag.